requirements.txt
python-telegram-bot[job-queue]==20.7
pytz==2023.3
python-dotenv==1.0.0
requests==2.31.0

